// src/api-swagger/models/FindOptionsQuery.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FindOptionsQuery = {
    page?: string;
    perPage?: string;
    searchValue?: string;
    searchFields?: Array<string>;
    additionalFields?: Array<string>;
    filters?: string;
};